Move recipe to fix reference from junit5.yml
