+++ src/main/java/org/openrewrite/java/testing/junit5/AddHamcrestJUnitDependency.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.testing.hamcrest;
+package org.openrewrite.java.testing.junit5;
 
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.ScanningRecipe;
